+++ react/Manager.js
@@ -4,6 +4,7 @@ import { orderFormConsumer, contextPropTypes } from 'vtex.store/OrderFormContext
 import ChangeAddressIcon from './components/ChangeAddressIcon'
 import ChangeAddressModal from './components/ChangeAddressModal'
 import AddressModal from './components/AddressModal'
+import hoistNonReactStatics from 'hoist-non-react-statics'
 import './global.css'
 
 /**
@@ -13,6 +14,8 @@ class AddressManager extends Component {
   static propTypes = {
     /* Context used to call address mutation and retrieve the orderForm */
     orderFormContext: contextPropTypes,
+    /* URL for the store logo */
+    logoUrl: PropTypes.string,
   }
 
   state = {
@@ -23,20 +26,21 @@ class AddressManager extends Component {
   handleCloseModal = () => this.setState({ isModalOpen: false })
 
   render() {
-    const { orderFormContext } = this.props
+    const { orderFormContext, logoUrl } = this.props
     const { shippingData } = orderFormContext.orderForm
 
     // If we don't know if there is an address or not we shouldn't load anything
     if (shippingData === undefined) return null
 
     // If there is no address, it means that the user isn't identified, and so the component won't render
-    if (!shippingData || !shippingData.address) return <AddressModal/>
+    if (!shippingData || !shippingData.address) return <AddressModal logoUrl={logoUrl}/>
 
     const { street, number } = shippingData.address
     const modalProps = {
       isOpen: this.state.isModalOpen,
       onClose: this.handleCloseModal,
-      orderFormContext: this.props.orderFormContext
+      orderFormContext: this.props.orderFormContext,
+      logoUrl,
     }
 
     return (
@@ -58,4 +62,19 @@ class AddressManager extends Component {
   }
 }
 
-export default orderFormConsumer(AddressManager)
+AddressManager.schema = {
+  title: 'address-locator.address-manager-title',
+  description: 'address-locator.address-manager-description',
+  type: 'object',
+  properties: {
+    logoUrl: {
+      type: 'string',
+      title: 'address-locator.address-manager.logo-title',
+      widget: {
+        'ui:widget': 'image-uploader'
+      }
+    },
+  }
+}
+
+export default hoistNonReactStatics(orderFormConsumer(AddressManager), AddressManager)